pyproject.toml
[tool.poetry]
package-mode=false

[tool.poetry.dependencies]
python="^3.11.9"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.35"}
grpcio = "^1.66.2"
grpcio-tools = "^1.66.2"
psycopg2-binary = "^2.9.9"
protobuf = "^5.28.2"
pip = "^24.2"
pydantic = "^2.9.2"
pydantic-settings = "^2.5.2"
alembic = "^1.13.3"
asyncpg = "^0.29.0"
black = "^24.8.0"
googleapis-common-protos = "^1.65.0"
grpcio-testing = "^1.66.2"
